Update exposure_to_chemical.tsv
Have not yet changed 'exposure to cadmium' to 'exposure to cadmium molecular entity' as Mondo is dependent on the term, does it matter if I change the term name if the ECTO ID and the CHEBI ref ID are the same? Same case for 'exposure to mercury'.
Other 'exposure to X' that should be 'exposure to X molecular entity' are changed and all duplicates removed.